Refactored VFS

+++ kernel/include/dev/uart.h
@@ -6,5 +6,5 @@ namespace uart
     void init();
     void write_char(const char c);
     void write_string(const char* string);
-    void write_number(const uint32_t number);
+    void write_number(const uint64_t number);
 }
